moved swap songs to PlaylistController, made test for moving songs in playlist
