+++ Controller/DbControllers/PlaylistController.cs
@@ -116,5 +116,17 @@ namespace Controller.DbControllers
         {
             return GetList().Where(x => x.ActivePlaylist).ToList();
         }
+        public void SwapSongs(int indexOne, int indexTwo, int playlistID)
+        {
+            var playlistSongController = PlaylistSongController.Create(new DatabaseContext());
+            var songOne = playlistSongController.GetPlaylistSongFromIndex(playlistID, indexOne);
+            var songTwo = playlistSongController.GetPlaylistSongFromIndex(playlistID, indexTwo);
+
+            songOne.Index = indexTwo;
+            songTwo.Index = indexOne;
+
+            playlistSongController.UpdatePlaylistSong(songOne);
+            playlistSongController.UpdatePlaylistSong(songTwo);
+        }
     }
 }
