+++ View/Screens/PlaylistScreen.xaml.cs
@@ -14,12 +14,14 @@ namespace View.Screens
 {
     partial class PlaylistScreen : ResourceDictionary
     {
+        private PlaylistController playlistController;
         public static PlaylistScreen Instance { get; set; }
 
         public PlaylistScreen()
         {
             this.InitializeComponent();
             Instance = this;
+            playlistController = PlaylistController.Create(new DatabaseContext());
         }
 
         public void OnNextSong(object sender, EventArgs e)
@@ -69,10 +71,9 @@ namespace View.Screens
         {
             var playlistID = (int)((Button)sender).Tag;
 
-            var playlistController = PlaylistController.Create(new DatabaseContext());
             var playlistName = playlistController.GetItem(playlistID).Name;
 
-            var removeConfirm = MessageBox.Show($"Are you sure you want to delete {playlistName}?", $"Remove {playlistName.ToString()}", MessageBoxButton.YesNo);
+            var removeConfirm = MessageBox.Show($"Are you sure you want to delete {playlistName}?", $"Remove {playlistName.ToString()}", MessageBoxButton.YesNoCancel);
             switch (removeConfirm)
             {
                 case MessageBoxResult.Yes:
@@ -80,6 +81,7 @@ namespace View.Screens
                     MainWindow.MenuItemRoutedEvent?.Invoke(this, new MenuItemRoutedEventArgs() { ScreenName = ScreenNames.PlaylistMenuScreen } );
                     break;
                 case MessageBoxResult.No:
+                case MessageBoxResult.Cancel:
                     break;
             }
         }
@@ -91,8 +93,8 @@ namespace View.Screens
 
             if (selectedSongInfo == null || selectedSongInfo.Index - 1 < 0) return;
 
-
-            SwapSongs(selectedSongInfo.Index, selectedSongInfo.Index - 1);
+            playlistController.SwapSongs(selectedSongInfo.Index, selectedSongInfo.Index - 1, MainWindow.CurrentPlayList.ID);
+            PlaylistDataContext.Instance.OnPropertyChanged("");
         }
 
         private void MoveDown_Click(object sender, RoutedEventArgs e)
@@ -102,24 +104,10 @@ namespace View.Screens
             var selectedSongInfo = (SongInfo)listView?.SelectedItem;
 
             if (selectedSongInfo == null || selectedSongInfo.Index + 1 >= listView.Items.Count) return;
-
-            SwapSongs(selectedSongInfo.Index, selectedSongInfo.Index + 1);
-        }
-
-        public void SwapSongs(int indexOne, int indexTwo)
-        {
-            var playlistSongController = PlaylistSongController.Create(new DatabaseContext());
-            int playlistID = MainWindow.CurrentPlayList.ID;
-            var songOne = playlistSongController.GetPlaylistSongFromIndex(playlistID, indexOne);
-            var songTwo = playlistSongController.GetPlaylistSongFromIndex(playlistID, indexTwo);
-
-            songOne.Index = indexTwo;
-            songTwo.Index = indexOne;
-
-            playlistSongController.UpdatePlaylistSong(songOne);
-            playlistSongController.UpdatePlaylistSong(songTwo);
-
+            
+            playlistController.SwapSongs(selectedSongInfo.Index, selectedSongInfo.Index + 1, MainWindow.CurrentPlayList.ID);
             PlaylistDataContext.Instance.OnPropertyChanged("");
         }
+
     }
 }
